Also factor TempData and HumidityData test classes into separate modules
