+++ lpdpower/testing/test_pscu_data.py
@@ -1,4 +1,4 @@
-"""Test TempData, HumidityData and PSCUData classes from lpdpower.
+"""Test PSCUData class from lpdpower.
 
 Tim Nicholls, STFC Application Engineering Group
 """
@@ -16,80 +16,7 @@ sys.modules['serial'] = Mock()
 sys.modules['Adafruit_BBIO'] = Mock()
 sys.modules['Adafruit_BBIO.GPIO'] = Mock()
 from odin.adapters.parameter_tree import ParameterAccessor
-from lpdpower.pscu_data import TempData, HumidityData, PSCUData, PSCUDataError
-
-class TestTempData():
-
-    @classmethod
-    @patch('lpdpower.i2c_device.smbus.SMBus')
-    def setup_class(cls, mock_bus):
-        cls.pscu = Mock()
-        cls.sensor_idx = 1
-        cls.temp_data = TempData(cls.pscu, cls.sensor_idx)
-
-    def test_setpoint_get(self):
-        self.temp_data.param_tree.get('setpoint')
-        self.pscu.get_temperature_set_point.assert_called_with(self.sensor_idx)
-
-    def test_temp_get(self):
-        self.temp_data.param_tree.get('temperature')
-        self.pscu.get_temperature.assert_called_with(self.sensor_idx)
-
-    def test_trace_get(self):
-        self.temp_data.param_tree.get('trace')
-        self.pscu.get_temperature_trace.assert_called_with(self.sensor_idx)
-
-    def test_tripped_get(self):
-        self.temp_data.param_tree.get('tripped')
-        self.pscu.get_temperature_tripped.assert_called_with(self.sensor_idx)
-
-    def test_disabled_get(self):
-        self.temp_data.param_tree.get('disabled')
-        self.pscu.get_temperature_disabled.assert_called_with(self.sensor_idx)
-
-    def test_name_get(self):
-        self.temp_data.param_tree.get('sensor_name')
-        self.pscu.get_temperature_name.assert_called_with(self.sensor_idx)
-
-    def test_mode_get(self):
-        self.temp_data.param_tree.get('mode')
-        self.pscu.get_temperature_mode.assert_called_with(self.sensor_idx)
-
-class TestHumidityData():
-
-    @classmethod
-    def setup_class(cls):
-        cls.pscu = Mock()
-        cls.sensor_idx = 1
-        cls.humidity_data = HumidityData(cls.pscu, cls.sensor_idx)
-
-    def test_setpoint_get(self):
-        self.humidity_data.param_tree.get('setpoint')
-        self.pscu.get_humidity_set_point.assert_called_with(self.sensor_idx)
-
-    def test_temp_get(self):
-        self.humidity_data.param_tree.get('humidity')
-        self.pscu.get_humidity.assert_called_with(self.sensor_idx)
-
-    def test_trace_get(self):
-        self.humidity_data.param_tree.get('trace')
-        self.pscu.get_humidity_trace.assert_called_with(self.sensor_idx)
-
-    def test_tripped_get(self):
-        self.humidity_data.param_tree.get('tripped')
-        self.pscu.get_humidity_tripped.assert_called_with(self.sensor_idx)
-
-    def test_disabled_get(self):
-        self.humidity_data.param_tree.get('disabled')
-        self.pscu.get_humidity_disabled.assert_called_with(self.sensor_idx)
-
-    def test_name_get(self):
-        self.humidity_data.param_tree.get('sensor_name')
-        self.pscu.get_humidity_name.assert_called_with(self.sensor_idx)
-
-    def test_mode_get(self):
-        self.humidity_data.param_tree.get('mode')
-        self.pscu.get_humidity_mode.assert_called_with(self.sensor_idx)
+from lpdpower.pscu_data import PSCUData, PSCUDataError
 
 
 class TestPscuData():
